PR/88735 is work in progress
When the criteria of the bug fix are loosen, ie a finalizer isn't a prerequisite,
I've a Linux specific failure with the test case 'defined_assignment_2.f90' which
has been introduced by the PR/46897. Currently, I can't debug in Linux. My bug fix
also seems to be somewhat fragile when a function result is involved in the RHS.